separado arquivos de middleware das rotas, feito validação extra na esclusão de incidents

+++ back-end/src/controllers/IncidentController.js
@@ -37,11 +37,18 @@ module.exports = {
         const { id } = req.params;
         const ong_id = req.headers.authorization;
 
+        const CheckID = await connection('incidents').where('id', id).first();
+        
+        if(!CheckID) {
+            return res.status(400).json({ error: "invalid ID"});
+        }
+
         const incident = await connection('incidents').where('id', id).select('ong_id').first();
 
         if(incident.ong_id !== ong_id ) {
             return res.status(401).json({ error: "Operation not permitted" });
         } 
+
         await connection('incidents').where('id', id).delete();
 
         return res.status(204).send();
